move backfill tolerance logic to timerangequery

+++ pkg/backends/options/options.go
@@ -84,7 +84,7 @@ type Options struct {
 	// BackfillTolerancePoints is similar to the MS version, except that it's final value is dependent
 	// on the query step value to determin the relative duration of backfill tolerance per-query
 	// When both are set, the higher of the two values is used
-	BackfillTolerancePoints int64 `yaml:"backfill_tolerance_points,omitempty"`
+	BackfillTolerancePoints int `yaml:"backfill_tolerance_points,omitempty"`
 	// PathList is a list of Path Options that control the behavior of the given paths when requested
 	Paths map[string]*po.Options `yaml:"paths,omitempty"`
 	// NegativeCacheName provides the name of the Negative Cache Config to be used by this Backend
